modified task 2

+++ Lesson_1/HOMEWORK/Task_2/Program.cs
@@ -2,9 +2,9 @@
 // два числа и выдаёт, какое число большее, а какое меньшее.
 
 Console.WriteLine("Введите первое число: ");
-double num1 = double.Parse(Console.ReadLine()!);
+int num1 = int.Parse(Console.ReadLine()!);
 Console.WriteLine("Введите второе число: ");
-double num2 = double.Parse(Console.ReadLine()!);
+int num2 = int.Parse(Console.ReadLine()!);
 
 if (num1 == num2)
 {
